+++ packages/agents-manager/src/components/zendesk-chat/style.scss
